this is the final update i think
new pics will be added in a folder. i was thinking of adding more features but then we wont be able to meet the deadline for the project submission. i hope this music player doesn't have any bugs. but if it has one pls write to me for the same. i know it lacks many common features, we had planned to add some extras too. so i'll work on it further after the end semester exams (my brain will work more cuz i ll have no tension of exams for a few days). i learnt alot from this project : )